Fixed a bug in version 1 (int8_t to int16_t) and in both versions added explicit float constants to stop warnings.

+++ The-ANSI-C-Programming-Language/Chapter01/fahrenheit-celcius-table-float.c
@@ -1,10 +1,12 @@
 #include <stdio.h>
+// import the new int types library
+#include <inttypes.h>
 
 // void here means don't take any arguments
 int main(void)
 {
 	float f, c;
-	int8_t lower, upper, increment;
+	int16_t lower, upper, increment;
 
 	lower = 0;
 	upper = 300;
@@ -14,9 +16,10 @@ int main(void)
 	printf("F\tC\n");
 	while (f <= upper) {
 		//c = 5 * (f - 32) / 9; // Integer division
-		c = (5.0 / 9.0) * (f - 32.0);
+		c = (5.0f / 9.0f) * (f - 32.0f);
 		//printf("%f\t%f\n", f, c); // not right justified
 		printf("%3.0f %6.1f\n", f, c); // right justified
 		f = f + increment;
+		//printf("f is %f\n", f);
 	}
 }
